Cleaning and refactoring #27
